+++ resources/views/internal/map/menu/businessLayers.blade.php
@@ -1,5 +1,3 @@
-@if ($mapView['controls']['layers'])
-    <fieldset class="daisy-kit-map__menu-section" data-daisy-kit-map-layers hidden>
-        <legend>{{ $mapView['labels']['businessLayers'] }}</legend>
-    </fieldset>
-@endif
+<fieldset class="daisy-kit-map__menu-section" data-daisy-kit-map-layers @disabled(! $control['enabled']) hidden>
+    <legend>{{ $control['label'] }}</legend>
+</fieldset>